Sensores/iot python fin/iot python fin/requirements.txt
pyserial
firebase-admin
paho-mqtt
